backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
sqlalchemy
python-dotenv
httpx
pytest
